GNOME Bugzilla #52890: replace time zone definitions when exporting events
There are reasons why calendars might contain events which
use an old and/or incomplete time zone definition:
* created with old Evolution release which did not yet use system time zones
* events created with other programs with TZIDs which could not be mapped yet
when the event was imported

This patch changes e_cal_get_timezone() so that if possible, the returned time
zone definition is taken from the system time zone database. Care was taken to
return the time zone with the TZID requested by the caller; otherwise the TZID
in a VEVENT would not match with the TZID in the system time zone.

I have checked a few situations and they all went through e_cal_get_timezone():
* exporting an event in SyncEvolution (via e_cal_component_as_string())
* saving an event in Evolution (also uses e_cal_component_as_string())
* sending an invitation in Evolution (via itip-util code)

The advantage of this patch is that updating system time zone information has
an effect on existing events. The drawback is that for historic events the
attached VTIMEZONE might be correct while the one generated by the current
Evolution for the current year is not. This drawback would go away as soon as
the generated VTIMEZONEs are complete. But until then the drawback of using an
outdated definition for a recurring event is probably worse, so the patch is
IMHO an improvement.

A (positive) side effect of the patch is that the backend is not called for
system time zones because they go through the new code path which takes the
definition from libical directly.

svn path=/branches/gnome-2-22/; revision=9025