#552261: sqlite is now a dependency of libcamel, and only of libcamel
Previously sqlite was listed as dependency of the various other
libs, but not of libcamel itself. It was missing in some places,
which caused compile errors. In the other places it unnecessarily
added libsqlite to the link line.

Together with Sri's fix of not exposing the sqlite handle to users
of libcamel unnecessarily, this commit fixes #552261.

svn path=/branches/gnome-2-24/; revision=9668