Bug 580021 — Port to external libgdata
Convert Google calendar and contacts backends to use the external libgdata,
depending on version >= 0.6.3. The e-d-s internal libgdata-1.2 has been
dropped completely.

The conversion of the Google calendar backend is untested and probably not
working, but since people should be using CalDAV to access their Google
Calendars, this shouldn't be a problem.

The conversion of the Google contacts backend is fairly well tested, and
has had support for a few new properties added. Most notably, addresses and
names are now stored on Google's servers in a structured format which more
closely maps to e-d-s' vCard representation than the previous flat string.

Closes: bgo#580021