BUGFIX 474948 (BNC): Expand <group> Inline Adds Control Codes.
* e-name-selector-entry.c (sanitize_string): Remove all the control
  characters from the headers. Doing this in sanitize string fixes it
  for all cases.

Bug Description:
If you auto-complete a Groupwise group into the To: field of an email messages,
you can right-mouse on it and select "Expand <group> Inline".  It should then
take all of the names and put them into the single line widget.  This works,
however it also adds control codes as well.