Fix leak in camel_folder_info_build()
All of the folders were added to a hash table with a non-strduped string as a
key, but then if the parent folder was missing, we created a new folderinfo for
the parent and inserted that into the hash table with a strduped key.  So when
the hash table was destroyed, the 'fake' parent folder keys were leaked.  To fix
this, I created the hash table with g_hash_table_new_full() which would g_free()
the keys and then always strduped the key.

https://bugzilla.gnome.org/show_bug.cgi?id=607588