Merge more cleanup bits from camel-gobject.
In particular, any method that takes a CamelException must return at
least a gboolean indicating success or failure.  This driven by the
convention in GLib of allowing GError arguments to be NULL, in which
case the function's return value is the only other way to report
failure.  GError will soon replace CamelException.