Bug 619347 - Return formatted address in e_destination_get_address
The documentation of this method claims to return the formatted name
yet it previously encoded them in quotable strings.  This patch fixes
this which results in strings containing non ASCII chracters to be
rendered correctly in the UI.  This also changes the vCard output but
vCard 3.0 specifies that quotable-printable inline ecnodings have been
eliminated.  In any case this quoting should be handled in
e_destination_export_to_vcard_attribute instead.