Handle [CLOSED] status (RFC5162).
This indicates precisely when the selected mailbox changes to the new
mailbox during the SELECT command, before its tagged completion.

This allows us to assign unsolicited FETCH responses to the correct
mailbox; with QRESYNC those will arrive before the SELECT completion.

It also theoretically allows us to pipeline SELECT requests, although
there's not a huge amount of point in that and we'll have other issues
with our implementation if we do that.