Propagate STATUS responses to folders directly from imapx_untagged()
Storing this in the server-global variables like is->exists is broken;
any SELECT command which is sent between the untagged STATUS response
and the tagged completion of the command is going to overwrite the data.
Not to mention the fact that LIST-STATUS and NOTIFY will be totally hosed,
when we get more than one untagged STATUS response between tagged commands.