Improve imapx command queue processing
We were breaking out of the loop which looked for commands to start on
the currently-selected folder, as soon as we reached any command which
was for any other folder.

This meant that we'd only ever run the _first_ command from the queue,
even if it was for another folder and there were other were other commands
with equal priority waiting to run on the _current_ folder. This leads to
a bunch of gratuitous ping-ponging between folders with SELECT, without
actually running any real commands between two consecutive SELECTs.