Simplify relationship between full and path names for imapx
There seems to be no point in the %25-style escapes, which we were applying
in full_to_path() only if dir_sep != '/', and then reversing in path_to_full()
unconditionally. Instead, just swap dir_sep with '/'. Aside from simplifying
the code, this also means that the path displayed to the user will be as close
as possible to the real name -- it means that my "be nasty to imapx" test
folder called fish"%sd is displayed properly, instead of as fish"%25sd.