Recalculate unread_on_server count after fetching flags.
Since we're not allowed to issue STATUS for the currently-selected mailbox,
the only way to ask the server for the count of unseen messages is to select
some _other_ mailbox and then issue STATUS. Even re-SELECTing the mailbox
doesn't give it, although a 'SEARCH UNSEEN' would do it, if we count the
responses.

But we _can_ work it out from the flags after we fetch them all.