Revamp imapx_job_refresh_info_start() to make use of modseq and uidnext
Also add the uidnext and modseq fields to CamelIMAPXSummary (from commit
86513477) although they're not going to be _stored_ yet (or potentially
ever in the 2.30 branch). But it's easier to keep the code the same.

This fixes the fact that for every STATUS request, we are first issuing
SELECT and then NOOP on the folder in question, which is _really_ slow on
some crappy servers like Microsoft Exchange. It was also ensuring that
_every_ STATUS request we were sending was invalid because it was for the
currently-selected folder (which is forbidden by RFC3501 §6.3.10.

(cherry-picked from commit b0214462be6e21c2a312d8fe3e31ffdd2825d153)