Remove courier-imap rename workaround
This is breaking things on non-buggy servers, because after renaming the
folder in question we also attempt to rename all its subfolders... and
quite reasonably get a failure.

Testing against current versions of Courier (4.4.0) seems to suggest
that it's not actually necessary, either -- perhaps the bug has been
fixed in the 9 years since this hack was first added to the imap back
end in commit bfce883e.