Fix nntp folder tree building
When it creates a 'virtual' NOSELECT|CHILDREN folder to fill a gap in
the hierarchy, it doesn't correctly insert that into the tree if it is a
top-level folder (i.e. last == NULL).

Once you fix that bug, you end up with two identical copies of the
tree-insertion code. I moved that out into a separate function.