Bug 622392 - Build failure due to missing nspr includes
camel.pc.in already defines @mozilla_nss@ and @mozilla_nspr@ in its
Requires list, but that only works for distros that ship their own nss
and nspr pkg-config files.

Debian, for example, doesn't (but it should).  So @mozilla_nss@ and
@mozilla_nspr@ are never defined.  Instead, the configure script falls
back to the older "manual" settings and defines MANUAL_NSS_CFLAGS and
MANUAL_NSS_LIBS.  So list those as well in camel.pc.in.  They will be
mutually exclusive with the pkg-config variables.

And please, for the love of free software, somebody push pkg-config
files to upstream nss/nspr so we can drop this gunk from configure.
It's 2010 already for cryin' out loud...