Keep better track of server-side UNREAD count.
There are various circumstances in which the UNREAD count will change --
new unseen messages arriving, flags changes from the client, unseen
messages being expunged, etc. Keep track of those to improve the
accuracy of the unread count, to avoid the QRESYNC sanity check
triggering so often on minor issues that weren't the bug I was looking
for. For this, we have to keep track of whether a FETCH or VANISHED response
was solicited or unsolicited in certain cases.
(cherry picked from commit 77b29a6e456f01e0a4241ca767233f77ab03a802)