Update unseen count as new messages come in
When we get notified of new messages in the idle thread, or when they arrive
during our re-fetch of all flags (and we suddenly find we've got flags for a
message that didn't previously exist), ensure that we update the unseen count
according to the flags of the new messages.

This _doesn't_ handle the case where we called fetch_new_messages() directly
from refresh_info_start() because we _knew_ there were new messages, but in
fact there are _more_ new messages than we expected. We may have to switch to
a model where we say 'update flags for any UID greater than $X'.
(cherry picked from commit f44ecb68679ea3134d7b3c5d5f346b640a11af45)