nntp: Fix two more 'returns FALSE on error' vs. 'returns 0 on success' errors
We really need to start being consistent about this.

Pass the GError pointer through to camel_folder_summary_save_to_db()
from nntp_folder_sync() too... although it seems to be NULL a lot of the
time anyway, but that's the fault of our caller. No need to compound it.