imapx: Ensure imapx_stream_fill() sets the error on zero-sized read.
We treat an empty read as an error, since we should never be here unless
we're either in the middle of reading a response, or poll() says the fd
is readable. Before commit 46938c348 (the conversion to GError), we'd set
a generic 'IO Error' exception in camel_imapx_stream_token() when
imapx_stream_fill() returned -1 to indicate an error. But now we don't
do that any more, so imapx_stream_fill() needs to reliably set the GError.

Otherwise we end up in an endless loop in imapx_step().