Refactor ECalBackendStore / ECalBackendFileStore.
These classes don't need to know the URI and source type of the backend,
they just need a file name.  Backends must now provide the name of the
cache file.  This breaks the API, but eliminates duplicate logic.

libedata-cal's shared object name is bumped for the API break.

Adapt backends to the new API.