Refactor ECalBackendCache.
ECalBackendCache doesn't need to know the URI of the backend, it just
needs a file name.  Backends must now provide the name of the cache
file.  This breaks the API, but eliminates duplicate logic.

Couple other miscellaneous functions removed:

    e_book_backend_construct()
    e_book_backend_cache_exists()
    e_book_backend_db_cache_exists()

libedata-book's shared object name is bumped for the API break.

Adapt backends to the new API.