Fix bogus console warning message from IMAP offline journal
There could be a deleted message in an offline journal in an IMAP account,
and while replying such journal the error message was reported to the caller
about not being able to find it, and the offline entry wasn't removed.
This change ignores the error to be able to get rid of the entry.