Bug #471079 (bnc) - Unable To Exclude Names
Unable To Exclude Names From Email Groups.

Bug Report:
I have created a personal "On This Computer" email group with multiple people.
When I type in that group name it auto-completes and displays in the To: field
underlined.

If you right mouse click on that group you see a list of all people that are in
the group with checkboxes next to their name.  You are supposed to be able to
uncheck people from the group and they will not receive the message.  However,
people that you uncheck still receive the message.