Fix failure to RSET SMTP connection before re-use.
We were only issuing a RSET in the case that we *didn't* need to, when
we'd sent a mail successfully. Fix that, so we send RSET after the RCPT
or DATA commands *fail*.

Do it by setting a 'need_rset' flag and then only actually sending RSET
if the connection really is going to be re-used. Most of the time, the
connection doesn't get re-used anyway so sending RSET would just be a
waste of time.

Mentioned in bug 630560.