Fix QRESYNC protocol violation (RH #648082)
The sequence-set '1:*' isn't valid in QRESYNC, and newer versions of
Dovecot (>=2.0) have started to object to it.

Ideally we'd omit the parameter completely, since it's supposed to be
optional — but *older* versions of Dovecot (<=1.2.13) require it.

We only really care about messages which we already knew about. For any
*new* messages, we're going to fetch the headers and the flags in an
explicit FETCH command anyway, and we'll ignore anything that QRESYNC
tells us about them. So just use the first and last UID from our existing
summary cache as the range for the QRESYNC command.

Fix a shadow declaration of the 'uid' variable in an inner block, while
we're at it.