bmo#6498 - Filters are not applied to imapx Inboxes
The default for the 'filter' option in imapx was set to False.
Now we make the default be True.  This means that incoming
messages through imapx will be filtered with Evolution's normal
filters.  A default of False means that your configuration doesn't
work - people have no idea why their filters aren't working.

Signed-off-by: Federico Mena Quintero <federico@novell.com>