Add versioning to D-Bus service names.
This is to address D-Bus interface mismatches when upgrading to and
running a new Evolution release while leaving the older E-D-S factory
services running, as many users do.

More details about the problem here:
http://mail.gnome.org/archives/evolution-hackers/2010-November/msg00041.html

Contrary to the mailing list post, I'm only adding version numbers to
the service names.  The new service names are:

    org.gnome.evolution.dataserver.AddressBook0
    org.gnome.evolution.dataserver.Calendar0

So, this works similar to shared library versioning.  If you add a D-Bus
method, remove a D-Bus method, change the signature of a D-Bus method,
or change the behavior of a D-Bus method in an incompatible way, bump
the version number at the end of the service name.

I'm also renaming our D-Bus object path and interface names to be less
repetitive, but this should be transparent given the new service names.