Remove ESoapMessage and ESoapResponse
These are only used by the EWS back end so far, and I want to revamp them
before we include them in a stable release. We have *large* SOAP requests
and responses, containing the full base64-encoded content of entire MIME
messages... we really want zero-copy for that part, and at the moment the
implementation has about three copies — we have to create a base64 string
in memory, pass it to e_soap_message_write_string() which makes a second
copy in the xmlDoc, then it's converted into a flattened buffer which is
passed to libsoup to send...

Just rip it all out for now; we'll keep a local copy in EWS and can add it
back into EDS 3.1 when we have it looking how we want it.

Don't bother with bumping the libedataserver soname for this; nobody was
using these anyway and they've never been in a stable release. Bumping
the soname as this stage of the release would be too painful.