Collect all test code into a top-level "tests" directory
This includes both automated unit tests and standalone demo programs.

Test code is partitioned by library.  So far we have:

    tests/libebook
    tests/libecal
    tests/libedata-cal
    tests/libedataserver
    tests/libedataserverui

This makes for a natural place to put additional E-D-S test code, but
still have it compartmentalized in one place (excluding Camel, which is
still tagging along with E-D-S but is not really part of E-D-S).