Replace CamelSessionThreadMsg with a more modern API.
This introduces camel_session_submit_job() to replace the older
CamelSessionThreadMsg API, providing a simpler mechanism for providers
to initiate low-priority background jobs.  Jobs can be submitted from
any thread, but execution of the job is always as follows:

1) A "job-started" signal is emitted from the thread in which the
   CamelSession was created.  This is typically the same thread that
   hosts the global default GMainContext, or "main" thread.

2) The callback function passed to camel_session_submit_job() is invoked
   from a different thread where it's safe to call synchronous functions.

3) Once the callback has returned, a "job-finished" signal is emitted
   from the same thread as "job-started" was emitted.

4) If a GDestroyNotify callback was passed to camel_session_submit_job(),
   it is invoked and passed the callback closure so it can be freed.

The signature of the callback function is easier to work with:

    void  (*CamelSessionCallback)  (CamelSession *session,
                                    GCancellable *cancellable,
                                    gpointer user_data,
                                    GError **error)

The cancellable is actually a CamelOperation, so status messages can be
pushed and popped and progress reported through the CamelOperation API.
The GError pointer is never NULL, so you can safely dereference it in
the callback function to check for errors.