Bug 628142 - Fix handling of simultaneous get_message requests
Drop the hash table of EFlags completely. It's broken, because the UID
we use as the hash key isn't actually unique; the same UID can exist in
multiple folders. And the lifetime issues on the EFlag weren't cleanly
solvable (yeah, we can add a refcounting wrapper, but ick).

We were *already* using imapx_is_job_in_queue() to check *properly* if
there was an existing fetch. So just implement a simple 'fetch counter'
with a GCond and a corresponding GMutex, bump that count by one each
time any fetch completes, and use the GCond when waiting for a *specific*
fetch to complete, inside a while (imapx_is_job_in_queue()) loop.