Bug 649433 - e-calendar-factory SIGSEGV while reading hash_table
The root cause of this issue was modules being loaded with RTLD_GLOBAL, and
having conflicting "global" symbols.

There is no reason for modules to be exporting their own symbols to the rest
of the process; add the G_MODULE_BIND_LOCAL flag to the g_module_open() call
to prevent that.