Camel: Fetch quota information asynchronously.
This bumps the libcamel soname again.

It turns out, camel_folder_get_quota_info() blocks.  So convert it to be
asynchronous like the other blocking functions, with corresponding class
methods:

  camel_folder_get_quota_info_sync()
  camel_folder_get_quota_info()
  camel_folder_get_quota_finish()

Also update the IMAP provider to make fetching IMAP quotas cancellable.
This is the only provider that currently implements this method.