Handle fields-of-interest for local addressbook backend.
This patch refactors the local addressbook backend to use
the new sqlitedb api instead of the old summary apis.
The result is that vcards are virtually built from the
sqlitedb cache when fields-of-interest is set.

Additionally, the patch adds the EBookBackend->notify_update()
vfunc which can be implemented for e_book_backend_notify_update().
The file backend uses this to notify with virtually created vcards
from the sqlite cache in response to contact additions and modifications.

This patch should address both bugs:
  https://bugzilla.gnome.org/show_bug.cgi?id=652179
  https://bugzilla.gnome.org/show_bug.cgi?id=652172