Add CamelNetworkService interface.
For now this interface is optional for network-based providers.
Eventually I'd like to move the connect() and disconnect() methods
from CamelService to CamelNetworkService entirely, but for now this
interface just collects duplicate code from the imap, imapx, nntp,
pop3 and smtp providers.  The connect_sync() method creates a TCP
stream and configures it to use a SOCKS proxy if available.