Bug 654480 - [imapx] Mark as Junk is not reliable
When comparing server-side message flags to local flags, disregard
non-permanent flags that appear to have been cleared on the server.

This solves the problem of IMAP servers not storing junk-related flags
permanently, then when we synchronize message flags with the server, our
local junk-related flags (which -are- stored permanently) get clobbered.