Be able to provide backtraces for e_ptr_tracker* functions
The e_ptr_tracker* functions from e-data-server-util.h track
pointers and reports those left in the queue at the end of
the application. With this change, when evolution-data-server
is configured with --enable-backtraces and required tools are
available, then also a backtrace when the pointer was added
to the queue is shown.

Thanks to Jan Kratochvil for his help with the dwfl part.