calendar file backend: fixed incomplete sanity check in e_cal_create_object()
e_cal_create_object() traditionally is used for a new events which are
unrelated to anything in the calendar. Adding detached recurrences
to an existing meeting series has to be done with e_cal_modify_object().

The code did not correctly reject the addition of a parent event for a
previously added child event event because lookup_component() returned
NULL in that case.

This commit renames lookup_component() and redefines the return value
to match what it is used for: checking for the existance of a UID.