Add various base classes for backends and servers.
EBackend is an abstract base class for address book and calendar
backends.

EBackendFactory is an abstract base class for creating address book and
calendar backends.  It is also an EExtension that extends EDataFactory.

EDBusServer is an abstract base class for an extensible D-Bus server.

EDataFactory is a type of EDBusServer that exports connection objects
for individual address books or calendars.