calendar file backend: revert adding EXDATE when creating detached recurrences
This reverts the part of 8631a8f which adds an EXDATE to the recurring
event when adding a detached recurrence. Adding that EXDATE is not
necessary, the detached recurrences overwrite the regular recurrences
based on UID/RECURRENCE-ID, and changes visible EDS behavior (breaks
SyncEvolution testing, for example).

As discussed in GNOME Bug #655253 and the "SyncEvolution + EClient API
+ EXDATE regression (Bug #655253)" mail thread, the real problem
seems to be related to UID handling in detached recurrences.