Bug #655253 - fix regression introduced with MOD_ONLY_THIS
Part of the code reorganization was removal of check_dup_uid() from
the normal "add component" code paths. It was only done as sanity
check when loading a file. That sanity check then incorrectly rewrote
UIDs, because it didn't take into account that detached recurrences
have (and must have) the same UID as the recurring master event.

This patch fixes check_dup_uid() accordingly. It moves the
uid g_free() into the bottom of the function to have a common
cleanup section.