addressbook: Optimizations to avoid vcard parsing
Now that vcards are lazily parsed, it makes sense to optimize
the rest of the code to avoid vcard parsing as much as possible.

In particular, it is now possible the pass the vcard UID when
constructing an EVCard object by using e_vcard_construct_with_uid().
This will avoid useless vcard parsing whenever getting the vcard
UID later. Moreover, setting the UID attribute on a EContact no
longer requires vcard parsing.