Provide CamelFolder::get_message_cached()
This is used in camel_folder_get_message_sync(), as the first attempt
to get a message without involving the folder lock, thus it's supposed
to be quicker. The implementer may not do any network I/O while inside
this function. Also, returning NULL is not considered as an error, it
just means that the message is not downloaded in the local cache yet,
thus the usual process of invoking folder's lock and calling actual
implementation of get_message_sync() follows. Defining this function
is not mandatory for descendants.