addressbook: Improve bulk contact removal code
The patch makes use of Berkeley DB transations in the file backend
and roll back the transaction in case of error to make the
behavior consistent with bulk addition.

This patch also makes the webdav backend behave as expected since
it does not support bulk removal and the implementation was broken.