Bug #661448 - e_util_copy_*_slist(): Avoid using append function
The current implementation calls g_slist_append() for each element
of the list to copy. However, g_slist_append() is expensive
because it iterate over the whole destination list.

This patch alters the behavior to first make a deep copy of the
list and then concatenate the destination list and the list copy.
The concatenation only iterates of the destination list once.